refactor structure - entities
