+++ src/entities/index.ts
@@ -1,0 +1,2 @@
+export { UserPayments } from './userPayments/UserPayments'
+export { UsersListTable } from './usersListTable/UsersListTable'
